configuration_1/backend/requirements.txt
torch==2.0.1
transformers==4.30.2
diffusers==0.19.3
accelerate==0.21.0
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
numpy==1.25.2
datasets==2.13.1
huggingface-hub==0.16.4
tqdm==4.66.1